refactor(model): move resolveEffectiveHelperModel out of @agent
CLI's enabled-models catalog imported @agent/runtime/helperModelName for a
pure validate-against-candidates helper, widening the host deep-import
ratchet. The function has no runtime dependency — house it under @model so
CLI and Settings share one copy without a new @agent edge.
